feature(ФормаОсновная): feature(ФормаОсновная): create music player
1.Music player - singleton;
2.Music player this class Player into player.js;
3.Music player
can
play,pause,next track, prev track;
4. Library into library.js now can't play, pause, next
track,
prev track;
5.Play music only with player.
6. Library not singleton.We can do large
count
librarys.
7. player.add(nameLibrary)- add all Tracks from library;
