+++ .eslintignore
@@ -1,4 +1,5 @@
 library.js
+player.js
 hw1
 hw2
 hw3
